# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Civic Intelligence — Citizen API
  version: 0.1.0
  description: |
    Evidence-grounded API for Kenyan legislation, regulations, and civic activity.

    **Architectural rule:** Every factual claim returned by the API carries
    citations. Responses that could not be fully verified are flagged.

  license: { name: MIT, url: https://github.com/Roy-Wanyoike/civic-intelligence/blob/main/LICENSE }
  contact: { url: https://github.com/Roy-Wanyoike/civic-intelligence }
servers:
  - url: http://localhost:9000/api/v1
    description: Local dev (BFF)
  - url: https://api.civicintelligence.dev/api/v1
    description: Production (placeholder)
...
paths:
...
  /scenarios:
    get:
      tags: [scenarios]
      summary: List scenarios
      description: |
        Every response carries an explicit `reality_layer` tag so HYPOTHETICAL
        outputs cannot be confused with observed civic facts.
      responses:
        '200':
          description: Scenarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  scenarios: { type: array, items: { $ref: '#/components/schemas/Scenario' } }
                  count: { type: integer }
                  disclaimer: { type: string }
    post:
      tags: [scenarios]
      summary: Create a scenario
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Scenario' }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/Scenario' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /scenarios/{id}:
    get:
      tags: [scenarios]
      summary: Get a scenario
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        '200':
          description: Scenario detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  scenario: { $ref: '#/components/schemas/Scenario' }
                  reality_layer: { type: string, enum: [FACT, OBSERVED, HYPOTHETICAL, MODELED, UNKNOWN] }
                  disclaimer: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [scenarios]
      summary: Validate / Run / Replay a scenario (selected via `action` query param)
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - in: query
          name: action
          required: true
          schema: { type: string, enum: [validate, run, replay] }
        - { in: query, name: seed, schema: { type: integer, description: 'Run reproducibility seed' } }
      responses:
        '200':
          description: Action result
          content:
            application/json:
              schema:
                type: object
                properties:
                  run: { $ref: '#/components/schemas/SimulationRun' }
                  reality_layer: { type: string, enum: [MODELED, HYPOTHETICAL] }
                  disclaimer: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC JWT issued by Keycloak (or any standards-compliant OIDC provider).
  schemas:
...
    Scenario:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        title: { type: string }
        description: { type: string }
        reality_layer: { type: string, enum: [FACT, OBSERVED, HYPOTHETICAL, MODELED, UNKNOWN] }
        scenario_type: { type: string }
        status: { type: string, enum: [DRAFT, READY, RUNNING, COMPLETED, FAILED, ARCHIVED] }
        model_id: { type: string }
        model_version: { type: string }
        baseline: { type: object }
        assumptions: { type: array, items: { $ref: '#/components/schemas/ScenarioAssumption' } }
        constraints: { type: array, items: { type: object } }
        variables: { type: array, items: { type: object } }
        time_horizon: { type: object }
        methodology: { $ref: '#/components/schemas/Methodology' }
        source_evidence: { type: array, items: { $ref: '#/components/schemas/EvidenceRef' } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ScenarioAssumption:
      type: object
      properties:
        id: { type: string }
        statement: { type: string }
        status: { type: string, enum: [PROPOSED, ACCEPTED, REJECTED, OBSERVED] }
        confidence: { type: string, enum: [high, medium, low, unknown] }
        source_evidence: { type: array, items: { $ref: '#/components/schemas/EvidenceRef' } }
    EvidenceRef:
      type: object
      properties:
        kind: { type: string, enum: [bill, act, regulation, hansard, committee_report, order_paper, gazette, news, journal] }
        ref_id: { type: string }
        url: { type: string }
        retrieved_at: { type: string, format: date-time }
    Methodology:
      type: object
      properties:
        approach: { type: string }
        data_sources: { type: array, items: { type: string } }
        limitations: { type: array, items: { type: string } }
        assumptions: { type: array, items: { type: string } }
        validation_status: { type: string, enum: [pending, passed, failed, partial] }
    SimulationRun:
      type: object
      properties:
        id: { type: string }
        scenario_id: { type: string }
        status: { type: string, enum: [PENDING, RUNNING, COMPLETED, FAILED, REPLAYED] }
        seed: { type: integer }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        duration: { type: string }
        result_summary: { type: object, nullable: true }
        reality_layer: { type: string, enum: [MODELED, HYPOTHETICAL] }
...
  responses:
...
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    BadRequest:
      description: Bad request — validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
security:
  - bearerAuth: []